move custom type validators to Types folder
