+++ src/Utils/index.js
@@ -1,5 +1,3 @@
 export { default as caseInsensitiveStringCompare } from './CaseInsensitiveStringCompare';
 export { default as getSpacingClasses } from './GetSpacingClasses';
 export { default as parseTextSize } from './ParseTextSize';
-export { default as boolRequiresOtherProp } from './BoolRequiresOtherProp';
-export { default as requiresOtherProp } from './RequiresOtherProp';
